Code Clean and Speed up

+++ src/com/muhavision/control/ExpoController.java
@@ -5,10 +5,11 @@ package com.muhavision.control;
  */
 public class ExpoController {
 
-    //If it crashes it's because of this + mami ne dotikaj se.
-    public static double getExpo(int power){
-        if (power<0) return-(0.6 * Math.exp(-0.035*(double)power) - 0.6);
-        return 0.6 * Math.exp(0.035*(double)power) - 0.6;
-    }
-    
+	// If it crashes it's because of this + mami ne dotikaj se.
+	public static double getExpo(int power) {
+		if (power < 0)
+			return -(0.6 * Math.exp(-0.035 * power) - 0.6);
+		return 0.6 * Math.exp(0.035 * power) - 0.6;
+	}
+
 }
